+++ src/main/java/com/perfumes/nuochoa/dto/ProductVariantDTO.java
@@ -1,13 +1,13 @@
 package com.perfumes.nuochoa.dto;
 
-import java.math.BigDecimal;
+
 
 public class ProductVariantDTO {
     private Long id;
     private String sku;
-    private BigDecimal volume;
-    private BigDecimal concentration; // 1 = EDT, 2 = EDP, 3 = Parfum
-    private BigDecimal price;
+    private Double volume;
+    private Double concentration; // 1 = EDT, 2 = EDP, 3 = Parfum
+    private Double price;
     private Integer stock;
 
     // Getters & Setters
@@ -17,14 +17,14 @@ public class ProductVariantDTO {
     public String getSku() { return sku; }
     public void setSku(String sku) { this.sku = sku; }
 
-    public BigDecimal getVolume() { return volume; }
-    public void setVolume(BigDecimal volume) { this.volume = volume; }
+    public Double getVolume() { return volume; }
+    public void setVolume(Double volume) { this.volume = volume; }
 
-    public BigDecimal getConcentration() { return concentration; }
-    public void setConcentration(BigDecimal concentration) { this.concentration = concentration; }
+    public Double getConcentration() { return concentration; }
+    public void setConcentration(Double concentration) { this.concentration = concentration; }
 
-    public BigDecimal getPrice() { return price; }
-    public void setPrice(BigDecimal price) { this.price = price; }
+    public Double getPrice() { return price; }
+    public void setPrice(Double price) { this.price = price; }
 
     public Integer getStock() { return stock; }
     public void setStock(Integer stock) { this.stock = stock; }
